Show directories with ending /, like ls(1)
This makes it more obvious when the leaf nodes are directories, for
example. I also refactored Root and Entry a bit:

They no longer keep a PathBuf stored, instead opting for storing just
the pre-calculated name. Root stores its own name which is the full
given path to keep the root output the same as before (but with a
trailing slash if it's a directory).